🌱 FEATURE | GCC -> CLANG | implement endswith function | int -> size_t or int -> ssize_t

+++ lib/phaseout/phase_append.c
@@ -9,15 +9,15 @@
 
 char *phase_append(char *base, char *add, char end)
 {
-    int base_length = phase_length(base, end);
-    int add_length = phase_length(add, end);
+    size_t base_length = phase_length(base, end);
+    size_t add_length = phase_length(add, end);
     char *array = phase_malloc((base_length + add_length), end);
-    int index = 0;
+    size_t index = 0;
 
-    for (int i = 0; i < base_length; i++, index++) {
+    for (size_t i = 0; i < base_length; i++, index++) {
         array[index] = base[i];
     }
-    for (int i = 0; i < add_length; i++, index++) {
+    for (size_t i = 0; i < add_length; i++, index++) {
         array[index] = add[i];
     }
