+++ lib/phaseout/phase_equals.c
@@ -7,13 +7,13 @@
 
 #include "PHASEOUT.h"
 
-int phase_equals(char *string, char *search, char end)
+size_t phase_equals(char *string, char *search, char end)
 {
-    int string_length = phase_length(string, end);
-    int search_length = phase_length(search, end);
+    size_t string_length = phase_length(string, end);
+    size_t search_length = phase_length(search, end);
 
     if (string_length == search_length && string_length > 0) {
-        for (int i = 0; i < string_length; i++) {
+        for (size_t i = 0; i < string_length; i++) {
             if (string[i] != search[i]) {
                 return (False);
             }
