+++ lib/phaseout/phase_prime.c
@@ -7,11 +7,11 @@
 
 #include "PHASEOUT.h"
 
-int phase_prime(int nb)
+size_t phase_prime(size_t nb)
 {
     if (nb <= 1)
         return (0);
-    for (int i = 2; i < nb; i++)
+    for (size_t i = 2; i < nb; i++)
         if (nb % i == 0)
             return (0);
     return (1);
