+++ lib/phaseout/phase_reverse.c
@@ -10,9 +10,9 @@
 char *phase_reverse(char *str, char end)
 {
     char temp = 0;
-    int size = phase_length(str, end);
+    size_t size = phase_length(str, end);
 
-    for (int i = 0; i < size / 2; i++) {
+    for (size_t i = 0; i < size / 2; i++) {
         temp = str[i];
         str[i] = str[size - 1 - i];
         str[size - 1 - i] = temp;
