+++ lib/phaseout/phase_writeline.c
@@ -7,7 +7,7 @@
 
 #include "PHASEOUT.h"
 
-int phase_writeline(char *str, char end)
+ssize_t phase_writeline(char *str, char end)
 {
     if (write(1, str, phase_length(str, end)))
         return (phase_writechar('\n'));
