+++ lib/phaseout/phase_writer.c
@@ -7,7 +7,7 @@
 
 #include "PHASEOUT.h"
 
-int phase_writer(char *str, char end, char last_character)
+ssize_t phase_writer(char *str, char end, char last_character)
 {
     if (write(1, str, phase_length(str, end)))
         return (phase_writechar(last_character));
